Initial commit of repository source base.  Directory structure is comprised of
"communication" -- those files related to client-server communication between
the iRobot on-board gumstix and other devices.

"roomba" -- those files related to controlling the gumstix which issues commands
and gathers sensor data from the iRobot.

"gumstix" -- those files which are used to configure the gumstix so that it may
successfully communicate with the iRobot.