Added more comprehensive unit test server. It creates a virtual perfect environment and receives commands from the Supervisor. It then puts together an appropriate sensor data string to send back to the Supervisor.
Changed the Supervisor to make a more educated choice on the command for the current
episde. It checks the score for each of the commands FORWARD, RIGHT, LEFT and of the
three matches it chooses the one closest to a goal, assuming a goal has been found.
Otherwise it talkes the option with the greatest score.

Changed tick method to be more obvious about what it's doing.
Added a createEpisode method to allocate space for episode and parse sensorData into it
Added a choseCommand method to determine a command to associate  with the episode