Have expanded the vectors of episodes and rules into vectors of vectors. This allows us to keep track of the layers of rules built from each other all in one overarching vector, and in the associated episodic memory we can keep track of the order in which we've seen the rules (by layer as well).
updateAllRules has been modified to be recursive and now takes a level (up to
MAX_META_DEPTH) that indexes into these overarching vectors. The process for creating the
rules is still the same, but now when a newrule is created and added a recursive call is
made that propogates this update through the layers of meta-rules and episodic memory.