I have added two functions.
The first is 'Rule* ruleMatch(int action)'
It basically takes the current state (defined by the current and immediate prior
episodes) and attempts to find a rule whose LHS matches this state.
If it finds such a rule, then it returns a pointer to it, and if no such match
is found then it returns NULL.

The second is 'int equalEpisodes(Episode* ep1, Episode* ep2)'
This is a helper method for above function. We had a compareEpisodes
function that returned a score representing how similar two episodes were. I
needed one that simply returned a boolean indicating a perfect match. So I put
this one together.