I have made a lot of changes.
I added a struct called Route. This is a global var that is the directions for the
Supervisor to follow.
It contains:
1. vector of pointers to ints that index into the rule list. This is the ordered list
			of Rules that the agent currently believes will take it to the goal.
			I chose to do pointers to ints instead of directly using pointers to rules
			because the Route changes often and it is easiest to delete the current
			route and start from scratch when recalculating.
2. int currRule, an index into route vector that gives the current Rule we're executing
3. int currEpInRule, an index into the Rule itself that we're currently executing
4. int needsRecalc, a flag that is flipped when the outcome of an action does not
			meet the expectation if TRUE then we need to recalculate the route
5. int numRules, number of rules in the route. I know that we could access this
			using the vector->size, but I thought that this would be quicker to access

Functions I added:

int planRoute(Episode* currEp) : takes an index to an episode with the current state
							   (sensor values) and determines a rule leading to a goal.
							   It then iterates through the list of rules and matches the
							   leftmost episode on the LHS against the current state, or the
							   outcome of a previous rule, depending on whether or not it's
							   the 1st rule in the route, in an attempt to bridge to the
							   goalRule.
							   It will always try to make the shortest route possible, using
							   the fewest rules possible.
							   It will eventually be more in depth and hopefully select Rules
							   intelligently. Currently it simply chooses any Rules that
							   match head-to-toe. Ideas on how to make this selection more
							   intelligent will be something we need to figure out. I
							   strongly feel that it is closely related to the idea of
							   remembering the swaps it makes, but we'll see.

void addRuleToRoute(int ruleIdx) : takes a Rule and adds it to the Route

int takeNextStep(Episode* currEp) : this method attempts to take the next step along the
								route. This is the function that determines if the outcomes
								are matching our expectation. It will continue to execute
								the Rule as long as everything continues to check out all
								right.

int setCommand2(Episode* ep) : this does the same thing as the original setCommand,
							   except it uses the Route and Rules to try and determine the
							   next action instead of the highest scoring match based on
							   NUM_TO_MATCH.

void displayRoute() : Prints out the route in reverse order with each rule on its own line.
					  This way you can read the route from bottom to top and it makes sense.


Currently there is a bug such that after the goal is found the first time it seg faults
within several episodes. I spent several hours fixing other small bugs that I thought were
the problems, but to no avail.

It is set up right now to choose random commands until after the first time it finds a goal.
After that it tries to set up a route and follow it until it's 'lost' at which point it
tries to determine a new route from its current state. I did this to check the route
following ability, but have not been able to check it due to the previously mentioned bug.

Part of the trouble I'm having with the debugging is that it seems that the delay between
printing to stdout and log files doesn't quite keep up with the running code. My reasoning is
that the seg fault doesn't always occur in the same place in the code. For awhile it happened
anywhere from 1 to 6 episodes after finding the goal the first time, most commonly between
1 and 3. Sometimes it would occur while printing out the rules, often times the last text
printed is inside the updateAllRules loop, but the server side shows several episodes to have
passed after the episode that is being processed on the screen.

That's the majority of what I've done so far. I may have missed a coupe of small changes I
made, but everything I mentioned above is the meat of my recent work.

																																																											HG: --
																																																																							HG: user: Zachary Faltersack <zachary.faltersack@gmail.com>
																																																																							HG: branch 'default'
																																																																							HG: changed communication/supervisorClient.c
																																																																							HG: changed supervisor/supervisor.c
																																																																							HG: changed supervisor/supervisor.h