I added some comments.
I also fixed the seg fault that was occuring before. Now the Supervisor
seg faults when it completes the final rule in the Route because I have
yet to add the check that takes the case of completing the final rule.

I began to change it back so that some of the time it chooses a random
command and some of the time it will try to follow the Route, this has
introduced some small bugs that result from not finishing the transition
process, but I have to go so I'll need to finish it later.

The process of choosing which Rules are used to construct a Route needs
to be better. The way it currently works doesn't lend itself to choosing
a very good route because it simply matches the first available Rules
end to front in the shortest way possible. Considering the states are
all so similar, this makes planRoute choose the same route most of the
time.
Again, I think that this is where that dictionary of replacements will
really come in handy, but we'll need more than that.