I cleaned up the unitTest files and split up the big function into smaller more manageable ones. I also spent some time making it a little more general by adding some #defines which should make it easier to implement different maps.
I also found a memory leak that was occuring because every time unitTest2
was called, memory was allocated to store the sensor string, but it was
never freed. So in the server code I made it free the memory after sending
the data to the Supervisor. Not sure if this will make a noticeable difference
in how it runs though.