I spent some time changing the unitTest files. I went ahead and changed it to read in the world from a file. If you look in world.maps there is a header that describes the format for creating new maps. It explains all the elements that can be in the world and how to represent them. the basic idea is that you can draw a map with a line above it that says:
mapNum mapWidth mapHeight

ie.

1 9 7
WWWWWWWWW
W>      W
W WW WW W
W WW WW W
W WW WW W
W      GW
WWWWWWWWW

2 5 3
WWWWW
W^ GW
WWWWW

.
.
.

So now when you call unitTestServer.out, you can pass it an argument that is the map
number. If no argument is passed then it defaults to 1, which is the world McKallum defined.

ie ./unitTestServer.out 1

During the course of testing I found some bugs that were introduced in the Supervisor
when we switched the Rule struct to have level instead of isBaseRule. When we did that,
we updated the calls to containsGoal to use level instead and in there we used !level to
determine if we were looking at episodes or rules. When we switched we ended up with some
calls that were getting the distinction backwards.

Also, the sequences after finding the first goal were only getting part of the action rules
added. This was because in tick, in order to speed things up I had made it skip updateRules
if the outcome matched our expectation. Unfortunately updateRules is where we add action rules
to sequences, so we only ever got the rules that did not match the expected outcome.