I have removed the superfluous functions from the Supervisor that will eventually become McCallum's Q-Learning agent.
In supervisor.h I've added some function declarations that I
believe will be needed in the future.

My plan is to implement a simple Heap ADT that will help the process
of locating the K nearest neighbors and make it so that the calculating
of the 'neighborhood metric' and the 'voting' can be done in one
fell swoop.