I decided to just move all the neighborhood stuff into the nsm files.
I copied the supervisorClient.c file and created a mccallumClient.c file
	with the appropriate changes to includes and updated variable
	names to match the newly defined episodes

I added a target in the makefile in the communication directory for mccClient.

Currently the mccallum code compiles. It is far from being able to run though.
No unit tests have yet been created to test functionality.

On the table: -	finish a couple things regarding the neighborhoods and their creation
	      -	work on algorithm that iterates through epmem and interacts with
			the neighborhood functions that have been defined
	      - once that is done, begin working on some unit tests that test the
			neighborhood functionality.

Once that is all done, the next steps will be to implement the more mathy-side
of things, especially regarding the Q-Learning aspect of this approach.