As far as I can tell I have the neighborhood parts working. I will continue to test but I wanted to get these updates pushed now that it's going.
Currently no decisions are made based on the neighborhoods. They are just created one
at a time and printed and then destroyed.

I added a lot of functions, but  they are pretty much all related to displaying the
neighborhoods and particular neighbors all for debugging purposes as well as to make
it more interesting to view the output.

I believe the next major steps are going to revolve around incorporating the Q-Learning
aspects of McCallum's approach.