Added "level" field to ReplacementStruct.  I believe that we will want to have replacements on each level of actions.
Updated initSupervisor() to include initilization of g_replacements.
Also
edited the code to be more stylish.

Update endSupervisor() to be more stylish, as well.  We still need code
here to clean up our replacements, but I'm not sure exactly how we'll
have replacements implemented yet.

Added prototypes and shells for doReplacement() and findReplacements().

I also noticed a potential memory leak in endSupervisor() under "clean
up episodes at this level," but it may just be a misunderstanding of
mine on how the episodes are laid out.  We might want to re-examine
that.