I changed the unit test to allow a map to have a single element added at a given point. A description on how this works is in the world.maps file.
In short, there are 4 optional items that define a:

mapElement which is the map item to add
goalCount which is the goal number that marks when the addition should happen
xCoord for new element
yCoord for new element

This could be useful for comparing how the different algorithms react to a
change in the environment. Especially since we've just been 'assuming' that
a dynamic world could hinder NSMQ and Ziggurat both.