I created 3 files: knearest.h, knearest.c, KNN_unitTest.c
knearest.h : This file contains the function declarations for knearest.c
knearest.c : This file contains the implementation of code that helps create and
		manage Neighborhoods.
KNN_unitTest.c : This is the unit test file to see if the code works. In addition,
		it provides an example of how to use the KNN mini-module.

I added a target to the makefile for KNN_unitTest.out.

I didn't notice that KNN had already been used as part of some file names.
I chose it for the obvious: K Nearest Neighbor.
I hope that that isn't too big of a deal, though it would be pretty easy
to refactor the code and remove it/change it.