Since my last commit I have made 3 significant changes:
First, Soar only now considers those commands that do not
send it into a wall.

Second, if once a best match has been chosen
findDiscountedCommandScore searches all the way to the current
episode for a reward, then a reward of 0 is returned.

Third, there is now a #define LOOK_AHEAD_N that when not set
to 0 will be a max limit on how far beyond the best match
findDiscountedCommandScore will search for a reward. If set to
0 then this is ignored (using a preprocessor #if statement).