I have fixed some bugs that cropped up.
I made some changes in initPlan regarding how
the offset was incorporated. There was a situation
where occasionally a plan would be initialized with
an invalid index into the first action. We would get a segfault
in that case.
I'm not entirely certain how or why it was happening, but I
fixed it by removing the check to see if it was a replan or not.
This affected how the offset was saved.
Additionally, if no previous offset calculated, then it would
be set to 1. I changed it to be set to 0, because in a 1 step
plan that was invalid.

I also kludged findInterimStartPartialMatch by adding a
static pointer to a Vector that indicates the last starting
vector that was returned. After selecting the new vector, it
is compared to the previous one and if they are the same,
then NULL is returned, forcing a random command.

This seems to work pretty well for now. I think this is
reasonable for now.