I added the ability to send a REFRESH command to the Eaters environment that refills it with food and moves the walls. (TODO: Allow a flag to use the same walls each time.)
I added some code that automatically chooses a random command
if the g_selfConfidence is below some number.
This helps prevent it from getting stuck against a wall.

Unfortunately, it seems as though this still yields a roughly
50% overall random commands versus 50% calculated commands.

But everything is running now and so as we work on finding
other bugs and fixing the algorithm we'll actually be able
to look at gathered data to check for increased performance.