I have added several functions to Ziggurat that help visualize a Route.
There are now a series of functions that can display an
EpisodeWME as if it were within an Eaters map.
You can also now display actions in a similar way.

What I'm really proud of is the function that composites
the actions within a sequence together to give a view of
all the area covered.
I was having trouble following the planning that was going
on because the string writeout of senses was so abstract.
I feel that this will help in the long run because now we can
visually compare where the agent thinks it is, and its plan,
with where it actually is in the environment.

I think a good step would be to go through a run and log
all of the sequences and actions that are generated and
confirm they are correct.
Also, we can use this to look into how the sequences are
being pieced together to create longer routes.

--------------------------------

I also added a #define to the WME code which turns on or
off the ability to only use walls when comparing episodes
and when running getNumMatches.

The results are much closer to what I was originally hoping
to see when this is turned on, but I know that it is not ideal
to use it. I just wanted to prove to myself that it really
would help.