Updated gatherInfo.sh to make it more robust.
Also, removed the temporary bit I added for Dustin and Ben that causes
the agent to make a random move from time to time.  Zach's low
confidence check is a better fix though still not my favorite.  In the
long run, I'd like to experiment again with confidence values on
Actions.