Cleaned up the simpleClient and simpleServer examples.  This pair of source, based largely on Beej's Network Programming Tutorial, are now fully commented to help newbies to the repository learn about socket programming.  The intention for this pair of files is to remain unaltered and exist only as a working example.
Altered the makefile for simpleServer so that it is not compiled statically.  Some network programming functionality, namely getaddrinfo(), requires runtime glibC libraries anyway.  These are available on the verdex pro default kernel images.  Tested new compilation of simpleServer.c on Bob + Frank, as compiled with the arm-linux-gnueabi-gcc compiler.

Added a note in nerves.c that it is an incomplete idea, a stepping stone towards a multi-threaded implementation of the roomba side of the testbed.