Affirmed the functionality of client.c + brainstem.c.  This pair of files creates a client-server relationship between the development machine and the gumstix, allowing one to manually control a roomba from the development machine's keyboard.  Rolled back the changes I made to move.c regarding usleep() vs sleep().  All move functionality now uses usleep().  Affirmed the changes were correct by testing with the drive.c and driveAndBump.c in roomba/examples.
Added a directory called "demo" which stores three binary files very useful to demonstrating the testbed to, say, prospective parents or students.  A README file in the directory explains how to run the demo code.

Removed main.c and makefile from the roomba/ directory, as this was redundant code -- a copy of driveAndBump.c in the examples.  Thus, there is only supporting functions in roomba/.  Actual programs are in roomba/examples or communication/.