- Corrected interface between mccallumClient and brainstem, supervisorClient and brainstem.  Previous implmentation of saccades was initializing the supervisor with too many possible commands to the roomba.  Thus, illegal commands were being sent to the roomba, and the roomba would stop executing commands after the first illegal command.  Failure took just a few seconds. - Removed all saccade implementation with SACC_FILTER preprocessor directives.  As a result, all source in communication is now compiling without failure from undefined CMD_SACC_X. - Removed all ugly debug output from brainstem.c.  It is now hidden beneath #ifdef DEBUG directives. - Corrected output from brainstem.c to print digits instead of characters.  Doi.
Executed brainstem.out on the Bob + Frank with two clients:

- ./mccallumClient.out 10.81.3.48 -c roomba
- ./supervisorClient.out 10.81.3.48 -c roomba

Both ran for 5+ minutes.  It is not clear to me if the clients are executing as intended -- I leave that up to Dr. Nuxoll's team -- but my interest is in preserving the interface between the roomba and the AI.  The strict alternation between command and sensor data is preserved after my summer code cleanup.